remove cartopy, geographiclib, mpi requirements for not; cast data to np.float; test now pass

+++ phasepapy/phasepicker/cf_aicd.py
@@ -9,7 +9,7 @@ class AicDeriv():
 
     def _statistics(self, optimized=True):
         npts = self.tr.stats.npts
-        data = self.tr.data;
+        data = self.tr.data.astype(np.float);
         # delta = 1.0/self.tr.stats.sampling_rate
 
         if (not optimized):
